refactor: 상세 화면 패키지 구조 개선 및 수신자 목록 UI 리팩터링
- `afternotedetail` 하위 패키지에 있던 상세 화면 관련 파일들을 `detail` 패키지로 이동하여 구조 단순화
- `ReceiversCard`에서 `InfoCard` 대신 `DetailSection`을 사용하도록 변경하여 상세 화면 레이아웃 일관성 확보
- 수신자 목록 아이템의 프로필 이미지 크기 조정(58dp -> 48dp) 및 텍스트 스타일 수정
- 수신자 관련 문자열 리소스 키 및 값을 `afternote_detail_extra_receivers_label`("추가 수신자")로 변경
- 패키지 위치 변경에 따른 임포트 경로 및 관련 내비게이션 코드 업데이트
