+++ feature/afternote/presentation/src/main/kotlin/com/afternote/feature/afternote/presentation/author/detail/AfternoteDetailState.kt
@@ -1,4 +1,4 @@
-package com.afternote.feature.afternote.presentation.author.detail.afternotedetail
+package com.afternote.feature.afternote.presentation.author.detail
 
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.Stable
