+++ feature/afternote/presentation/src/main/kotlin/com/afternote/feature/afternote/presentation/author/detail/SocialNetworkDetailScreen.kt
@@ -1,5 +1,6 @@
-package com.afternote.feature.afternote.presentation.author.detail.afternotedetail
+package com.afternote.feature.afternote.presentation.author.detail
 
+import androidx.compose.foundation.clickable
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Column
@@ -10,13 +11,16 @@ import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.layout.size
 import androidx.compose.foundation.rememberScrollState
 import androidx.compose.foundation.verticalScroll
+import androidx.compose.material3.HorizontalDivider
 import androidx.compose.material3.Icon
 import androidx.compose.material3.IconButton
 import androidx.compose.material3.Scaffold
+import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
-import androidx.compose.runtime.Immutable
 import androidx.compose.runtime.getValue
+import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.remember
+import androidx.compose.runtime.setValue
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.res.painterResource
@@ -26,11 +30,11 @@ import androidx.compose.ui.unit.dp
 import androidx.hilt.lifecycle.viewmodel.compose.hiltViewModel
 import androidx.lifecycle.compose.collectAsStateWithLifecycle
 import com.afternote.core.ui.badge.RecipientDesignationBadgeState
+import com.afternote.core.ui.theme.AfternoteDesign
 import com.afternote.core.ui.theme.AfternoteTheme
 import com.afternote.core.ui.topbar.DetailTopBar
 import com.afternote.feature.afternote.domain.AfternoteServiceType
 import com.afternote.feature.afternote.presentation.R
-import com.afternote.feature.afternote.presentation.author.detail.AfternoteDetailViewModel
 import com.afternote.feature.afternote.presentation.author.detail.model.AfternoteDetailEvent
 import com.afternote.feature.afternote.presentation.author.detail.model.AfternoteDetailUiState
 import com.afternote.feature.afternote.presentation.author.navigation.DesignPendingDetailContent
@@ -38,21 +42,27 @@ import com.afternote.feature.afternote.presentation.author.navigation.DetailLoad
 import com.afternote.feature.afternote.presentation.author.navigation.HandleDeleteResult
 import com.afternote.feature.afternote.presentation.shared.detail.AfternoteDetailServiceHeaderWithRecipientChip
 import com.afternote.feature.afternote.presentation.shared.detail.DeleteConfirmDialog
+import com.afternote.feature.afternote.presentation.shared.detail.DetailInfoRow
+import com.afternote.feature.afternote.presentation.shared.detail.DetailSection
 import com.afternote.feature.afternote.presentation.shared.detail.EditDropdownMenu
 import com.afternote.feature.afternote.presentation.shared.detail.MessageSection
 import com.afternote.feature.afternote.presentation.shared.detail.ProcessingMethodsSection
-import com.afternote.feature.afternote.presentation.shared.detail.ReceiversCard
 import com.afternote.feature.afternote.presentation.shared.model.ReceiverUiModel
 import com.afternote.feature.afternote.presentation.shared.util.getIconResForServiceName
 
 /**
- * 갤러리 상세 Stateful Route.
+ * 소셜 네트워크 상세 Stateful Route.
  *
- * [SocialNetworkDetailRoute] 와 동일하게 공용 [AfternoteDetailViewModel]·[AfternoteDetailUiState] 를 쓰고,
- * 성공 시 [AfternoteServiceType.GALLERY_AND_FILES] 타입을 검사해 폴백한다.
+ * Now in Android 가이드의 Route + Screen 분리 패턴을 따른다.
+ * - 상세/작성자/삭제 상태는 공용 [AfternoteDetailViewModel] 에서 관리한다
+ *   (상세 목적지마다 별도의 백스택 엔트리이므로 VM 인스턴스는 화면마다 갈리지만 클래스는 동일).
+ *   초기 상세 로드는 ViewModel 의 `init` 과 네비게이션 인자(`itemId`)로만 트리거한다.
+ * - UI 는 [SocialNetworkDetailScreen] (Stateless) 에 위임한다.
+ * - 방어적으로 `detail.type` 을 검사해 SOCIAL_NETWORK 가 아닌 타입이 들어오면
+ *   [DesignPendingDetailContent] 로 폴백한다.
  */
 @Composable
-internal fun GalleryDetailRoute(
+internal fun SocialNetworkDetailRoute(
     onBack: () -> Unit,
     onNavigateToEditor: (itemId: String) -> Unit,
     viewModel: AfternoteDetailViewModel = hiltViewModel(),
@@ -76,14 +86,21 @@ internal fun GalleryDetailRoute(
             )
 
             val detail = state.detail
-            if (detail.type != AfternoteServiceType.GALLERY_AND_FILES) {
+            if (detail.type != AfternoteServiceType.SOCIAL_NETWORK) {
                 DesignPendingDetailContent(onBackClick = onBack)
             } else {
-                GalleryDetailScreen(
+                val method = detail.processing?.method ?: ""
+
+                SocialNetworkDetailScreen(
                     content =
-                        GalleryDetailContent(
+                        SocialNetworkDetailContent(
                             serviceName = detail.title,
                             userName = state.authorDisplayName,
+                            accountId = detail.credentials?.id ?: "",
+                            password = detail.credentials?.password ?: "",
+                            accountProcessingMethod = method,
+                            processingMethods = detail.processing?.actions ?: emptyList(),
+                            message = detail.processing?.leaveMessage ?: "",
                             finalWriteDate = detail.timestamps.updatedAt.ifEmpty { detail.timestamps.createdAt },
                             afternoteEditReceivers =
                                 detail.receivers.map { r ->
@@ -93,8 +110,7 @@ internal fun GalleryDetailRoute(
                                         label = r.relation,
                                     )
                                 },
-                            processingMethods = detail.processing?.actions ?: emptyList(),
-                            message = detail.processing?.leaveMessage ?: "",
+                            iconResId = getIconResForServiceName(detail.title),
                         ),
                     onBackClick = onBack,
                     onEditClick = { onNavigateToEditor(detail.id.toString()) },
@@ -106,28 +122,13 @@ internal fun GalleryDetailRoute(
 }
 
 /**
- * 갤러리 상세 표시 데이터.
- */
-@Immutable
-data class GalleryDetailContent(
-    val serviceName: String = "",
-    val userName: String = "",
-    val finalWriteDate: String = "",
-    val afternoteEditReceivers: List<ReceiverUiModel> = emptyList(),
-    val processingMethods: List<String> = emptyList(),
-    val message: String = "",
-)
-
-/**
- * 갤러리 애프터노트 상세 화면 (Stateless).
- *
- * [SocialNetworkDetailScreen] 과 동일한 Scaffold·TopBar·드롭다운·스크롤 modifier 패턴을 따른다.
+ * 소셜 네트워크 애프터노트 상세 화면.
  */
 @Composable
-fun GalleryDetailScreen(
+fun SocialNetworkDetailScreen(
     onBackClick: () -> Unit,
     modifier: Modifier = Modifier,
-    content: GalleryDetailContent = GalleryDetailContent(),
+    content: SocialNetworkDetailContent = SocialNetworkDetailContent(),
     isEditable: Boolean = true,
     onEditClick: () -> Unit = {},
     onDeleteConfirm: () -> Unit = {},
@@ -173,7 +174,7 @@ fun GalleryDetailScreen(
             )
         },
     ) { paddingValues ->
-        GalleryDetailScrollContent(
+        SocialNetworkDetailScrollContent(
             content = content,
             modifier =
                 Modifier
@@ -186,8 +187,8 @@ fun GalleryDetailScreen(
 // region — Scroll content
 
 @Composable
-private fun GalleryDetailScrollContent(
-    content: GalleryDetailContent,
+private fun SocialNetworkDetailScrollContent(
+    content: SocialNetworkDetailContent,
     modifier: Modifier = Modifier,
 ) {
     Column(
@@ -199,7 +200,7 @@ private fun GalleryDetailScrollContent(
                 .padding(horizontal = 20.dp),
     ) {
         AfternoteDetailServiceHeaderWithRecipientChip(
-            iconResId = getIconResForServiceName(content.serviceName),
+            iconResId = content.iconResId,
             serviceName = content.serviceName,
             finalWriteDate = content.finalWriteDate,
             recipientBadgeState =
@@ -212,43 +213,116 @@ private fun GalleryDetailScrollContent(
 
         Spacer(modifier = Modifier.height(31.dp))
         Column(verticalArrangement = Arrangement.spacedBy(16.dp)) {
-            ReceiversCard(receivers = content.afternoteEditReceivers)
+            AccountSection(
+                accountId = content.accountId,
+                password = content.password,
+            )
             ProcessingMethodsSection(methods = content.processingMethods)
             MessageSection(message = content.message)
         }
     }
 }
 
+/**
+ * 소셜 네트워크 상세 전용 ACCOUNT(아이디·비밀번호) 섹션.
+ *
+ * 섹션 뼈대는 공용 [DetailSection], 행 레이아웃은 [DetailInfoRow] 를 쓰고,
+ * 비밀번호 표시 토글 상태·동작만 이 블록에 둔다.
+ */
+@Composable
+private fun AccountSection(
+    accountId: String,
+    password: String,
+    modifier: Modifier = Modifier,
+) {
+    var passwordVisible by remember { mutableStateOf(false) }
+
+    DetailSection(
+        iconResId = com.afternote.core.ui.R.drawable.core_ui_user,
+        label = stringResource(R.string.feature_afternote_detail_section_account),
+        modifier = modifier,
+    ) {
+        Column(verticalArrangement = Arrangement.spacedBy(8.dp)) {
+            DetailInfoRow(
+                iconResId = com.afternote.core.ui.R.drawable.core_ui_user,
+                label = stringResource(R.string.feature_afternote_detail_label_id),
+                value = accountId,
+            )
+            HorizontalDivider(
+                color = AfternoteDesign.colors.gray2,
+                thickness = 1.dp,
+            )
+            DetailInfoRow(
+                iconResId = com.afternote.core.ui.R.drawable.core_ui_ic_deep_thought_moon,
+                label = stringResource(R.string.feature_afternote_detail_label_password),
+                value =
+                    if (passwordVisible) {
+                        password
+                    } else {
+                        stringResource(R.string.feature_afternote_detail_password_mask)
+                    },
+                trailingContent = {
+                    Text(
+                        text =
+                            if (passwordVisible) {
+                                stringResource(R.string.feature_afternote_detail_password_hide)
+                            } else {
+                                stringResource(R.string.feature_afternote_detail_password_show)
+                            },
+                        style = AfternoteDesign.typography.captionLargeR,
+                        color = AfternoteDesign.colors.b1,
+                        modifier =
+                            Modifier.clickable {
+                                passwordVisible = !passwordVisible
+                            },
+                    )
+                },
+            )
+        }
+    }
+}
+
 // endregion
 
-internal val GALLERY_PREVIEW_CONTENT =
-    GalleryDetailContent(
-        serviceName = "갤러리",
+/**
+ * Android Studio @Preview 전용. Debug 빌드 런타임 목업은
+ * `AfternoteDebugMockNetworkInterceptor`의 GET `/api/afternotes/{id}` (예: id 1, 3)를 사용한다.
+ */
+private val PreviewSocialNetworkInstaContent =
+    SocialNetworkDetailContent(
+        serviceName = "인스타그램",
         userName = "서영",
+        accountId = "qwerty123",
+        password = "qwerty123",
+        accountProcessingMethod = "MEMORIAL",
+        processingMethods = listOf("게시물 내리기", "추모 게시물 올리기", "추모 계정으로 전환하기"),
+        message = "이 계정에는 우리 가족 여행 사진이 많아.\n계정 삭제하지 말고 꼭 추모 계정으로 남겨줘!",
         finalWriteDate = "2025.11.26",
-        processingMethods = listOf("'엽사' 폴더 박선호에게 전송", "'흑역사' 폴더 삭제"),
         afternoteEditReceivers =
             listOf(
-                ReceiverUiModel(id = "1", name = "김지은", label = "친구"),
-                ReceiverUiModel(id = "2", name = "김혜성", label = "친구"),
+                ReceiverUiModel(
+                    id = "1",
+                    name = "황규운",
+                    label = "친구",
+                ),
             ),
     )
 
 @Preview(showBackground = true)
 @Composable
-private fun GalleryDetailScreenPreview() {
+private fun SocialNetworkDetailScreenPreview() {
     AfternoteTheme {
-        GalleryDetailScreen(
-            content = GALLERY_PREVIEW_CONTENT,
+        SocialNetworkDetailScreen(
+            content = PreviewSocialNetworkInstaContent,
             onBackClick = {},
             onEditClick = {},
         )
     }
 }
 
-@Preview(showBackground = true, name = "Edit Dropdown Menu")
+@Preview(showBackground = true)
 @Composable
-private fun GalleryDetailScreenWithDropdownPreview() {
+private fun SocialNetworkDetailScreenWithDropdownPreview() {
     AfternoteTheme {
         val stateWithDropdown =
             remember {
@@ -256,44 +330,11 @@ private fun GalleryDetailScreenWithDropdownPreview() {
                     toggleDropdownMenu()
                 }
             }
-        GalleryDetailScreen(
-            content = GALLERY_PREVIEW_CONTENT,
+        SocialNetworkDetailScreen(
+            content = PreviewSocialNetworkInstaContent,
             onBackClick = {},
             onEditClick = {},
             state = stateWithDropdown,
-        )
-    }
-}
-
-@Preview(showBackground = true)
-@Composable
-private fun GalleryDetailScreenWithDeleteDialogPreview() {
-    AfternoteTheme {
-        val stateWithDialog =
-            remember {
-                AfternoteDetailState().apply {
-                    showDeleteDialog()
-                }
-            }
-        GalleryDetailScreen(
-            content = GALLERY_PREVIEW_CONTENT,
-            onBackClick = {},
-            onEditClick = {},
-            state = stateWithDialog,
-        )
-    }
-}
-
-@Preview(showBackground = true)
-@Composable
-private fun GalleryDetailScreenReceiverModePreview() {
-    AfternoteTheme {
-        GalleryDetailScreen(
-            content = GALLERY_PREVIEW_CONTENT,
-            isEditable = false,
-            onBackClick = {},
-            onEditClick = {},
-            state = rememberAfternoteDetailState(),
         )
     }
 }
